# go/little-gin-admin/server/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample Server pets",
    "title": "Swagger Example API",
    "contact": {},
    "license": {
      "name": "MIT"
    },
    "version": "0.0.1"
  },
  "basePath": "/",
  "paths": {
...
    "/fileUploadAndDownload/deleteFile": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "ExaFileUploadAndDownload"
        ],
        "summary": "删除文件",
        "parameters": [
          {
            "description": "传入文件里面id即可",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/model.ExaFileUploadAndDownload"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{\"success\":true,\"data\":{},\"msg\":\"返回成功\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "/fileUploadAndDownload/getFileList": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "ExaFileUploadAndDownload"
        ],
        "summary": "分页文件列表",
        "parameters": [
          {
            "description": "分页获取文件户列表",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/request.PageInfo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{\"success\":true,\"data\":{},\"msg\":\"获取成功\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "/fileUploadAndDownload/upload": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "ExaFileUploadAndDownload"
        ],
        "summary": "上传文件示例",
        "parameters": [
          {
            "type": "file",
            "description": "上传文件示例",
            "name": "file",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{\"success\":true,\"data\":{},\"msg\":\"上传成功\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "/sysDictionaryDetail/createSysDictionaryDetail": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "SysDictionaryDetail"
        ],
        "summary": "创建SysDictionaryDetail",
        "parameters": [
          {
            "description": "创建SysDictionaryDetail",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/model.SysDictionaryDetail"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{\"success\":true,\"data\":{},\"msg\":\"获取成功\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/sysDictionaryDetail/deleteSysDictionaryDetail": {
      "delete": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "SysDictionaryDetail"
        ],
        "summary": "删除SysDictionaryDetail",
        "parameters": [
          {
            "description": "删除SysDictionaryDetail",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/model.SysDictionaryDetail"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{\"success\":true,\"data\":{},\"msg\":\"删除成功\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/sysDictionaryDetail/findSysDictionaryDetail": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "SysDictionaryDetail"
        ],
        "summary": "用id查询SysDictionaryDetail",
        "parameters": [
          {
            "description": "用id查询SysDictionaryDetail",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/model.SysDictionaryDetail"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{\"success\":true,\"data\":{},\"msg\":\"查询成功\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/sysDictionaryDetail/getSysDictionaryDetailList": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "SysDictionaryDetail"
        ],
        "summary": "分页获取SysDictionaryDetail列表",
        "parameters": [
          {
            "description": "分页获取SysDictionaryDetail列表",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/request.SysDictionaryDetailSearch"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{\"success\":true,\"data\":{},\"msg\":\"获取成功\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/sysDictionaryDetail/updateSysDictionaryDetail": {
      "put": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "SysDictionaryDetail"
        ],
        "summary": "更新SysDictionaryDetail",
        "parameters": [
          {
            "description": "更新SysDictionaryDetail",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/model.SysDictionaryDetail"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{\"success\":true,\"data\":{},\"msg\":\"更新成功\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "model.ExaFileUploadAndDownload": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "tag": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }
      }
    },
...
    "model.SysDictionaryDetail": {
      "type": "object",
      "properties": {
        "label": {
          "type": "string"
        },
        "sort": {
          "type": "integer"
        },
        "status": {
          "type": "boolean"
        },
        "sysDictionaryID": {
          "type": "integer"
        },
        "value": {
          "type": "integer"
        }
      }
    },
...
    "request.PageInfo": {
      "type": "object",
      "properties": {
        "page": {
          "type": "integer"
        },
        "pageSize": {
          "type": "integer"
        }
      }
    },
...
    "request.SysDictionaryDetailSearch": {
      "type": "object",
      "properties": {
        "label": {
          "type": "string"
        },
        "page": {
          "type": "integer"
        },
        "pageSize": {
          "type": "integer"
        },
        "sort": {
          "type": "integer"
        },
        "status": {
          "type": "boolean"
        },
        "sysDictionaryID": {
          "type": "integer"
        },
        "value": {
          "type": "integer"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "x-token",
      "in": "header"
    }
  }
}